backend/requirements.txt
asgiref==3.5.2
backports.zoneinfo==0.2.1
Django==4.1.3
django-admin-volt==1.0.4
djangorestframework==3.14.0
numpy==1.23.5
pytz==2022.6
scipy==1.9.3
sqlparse==0.4.3
